better support cuda32

+++ src/cuckatoo/mean.cu
@@ -54,11 +54,11 @@ const u32 ZMASK     = NZ - 1;
 #endif
 #define NEPS 128
 
-const u32 EDGES_A = NZ * NEPS_A / NEPS;
-const u32 EDGES_B = NZ * NEPS_B / NEPS;
+const u64 EDGES_A = NZ * NEPS_A / NEPS;
+const u64 EDGES_B = NZ * NEPS_B / NEPS;
 
-const u32 ROW_EDGES_A = EDGES_A * NY;
-const u32 ROW_EDGES_B = EDGES_B * NY;
+const u64 ROW_EDGES_A = EDGES_A * NY;
+const u64 ROW_EDGES_B = EDGES_B * NY;
 
 // Number of rows in bufferB not overlapping bufferA
 #ifndef NRB1
@@ -871,12 +871,12 @@ CALL_CONVENTION void fill_default_params(SolverParams* params) {
   params->device = 0;
   params->ntrims = tp.ntrims;
   params->expand = tp.expand;
-  params->genablocks = min(tp.genA.blocks, NEDGES/tp.genA.tpb);
+  params->genablocks = min(tp.genA.blocks, (u16)(NEDGES/tp.genA.tpb));
   params->genatpb = tp.genA.tpb;
   params->genbtpb = tp.genB.tpb;
   params->trimtpb = tp.trim.tpb;
   params->tailtpb = tp.tail.tpb;
-  params->recoverblocks = min(tp.recover.blocks, NEDGES/tp.recover.tpb);
+  params->recoverblocks = min(tp.recover.blocks, (u16)(NEDGES/tp.recover.tpb));
   params->recovertpb = tp.recover.tpb;
   params->cpuload = false;
 }
